+++ lib/attr_cipher/attr_cipher.rb
@@ -1,6 +1,5 @@
 require 'active_record'
 require 'active_support/all'
-require 'yaml'
 
 module AttrCipher
   extend ActiveSupport::Concern
@@ -30,16 +29,14 @@ module AttrCipher
       define_method attribute do
         value = instance_variable_get("@#{attribute}")
         cipher_value = send("#{attribute}_cipher") unless value
-        value = cipher.decrypt(secret, cipher_value) if cipher_value
-        value = YAML::load(value) if serialize
+        value = cipher.decrypt(secret, cipher_value, serialize) if cipher_value
         instance_variable_set("@#{attribute}", value)
       end
 
       define_method "#{attribute}=" do |value|
         instance_variable_set("@#{attribute}", value)
-        value = YAML::dump(value) if serialize
         send("#{attribute}_cipher=", nil)
-        send("#{attribute}_cipher=", cipher.encrypt(secret, value)) if value && value != ""
+        send("#{attribute}_cipher=", cipher.encrypt(secret, value, serialize)) if value && value != ""
       end
     end
   end