+++ spec/support/custom_cipher.rb
@@ -1,8 +1,8 @@
 module CustomCipher
-  def self.encrypt(secret, value)
+  def self.encrypt(secret, value, serialize = false)
     value.to_s.reverse
   end
-  def self.decrypt(secret, value)
+  def self.decrypt(secret, value, serialize = false)
     value.to_s.reverse
   end
 end